Move the sync methods onto a `PathSync` class.
The ambient augmentation is not great in JS, because a single import (even in a dependency) pulls in the entire functionality. I'd love something like Rust where you can't use the sync methods without explicitly importing a trait in the file that uses is, but that's not really an option.